perf: 메인 페이지 TTFB 최적화 - auth 중복 호출 제거 및 SSR prefetch 전환
