+++ package.json
@@ -37,6 +37,7 @@
     "@supabase-cache-helpers/postgrest-react-query": "^1.11.4",
     "@supabase/ssr": "^0.5.2",
     "@supabase/supabase-js": "^2.47.12",
+    "@suspensive/react": "^3.20.3",
     "@tailwindcss/postcss": "^4.2.1",
     "@tanstack/react-query": "^5.56.2",
     "@tanstack/react-table": "^8.20.5",
