+++ src/actions/program-action.ts
@@ -244,18 +244,23 @@ export async function getAllJudgingWorkspaces(
   searchKeyword?: string,
   isParticipating?: boolean,
   exactJudgingRoundId?: string,
-  statusFilter?: string
+  statusFilter?: string,
+  preAuthedUserId?: string
 ): Promise<AllJudgingWorkspacesResult> {
   const supabase = await createClient();
 
-  const {
-    data: { user },
-    error: authError,
-  } = await supabase.auth.getUser();
-  const userId = user?.id;
+  let userId = preAuthedUserId;
 
-  if (authError || !userId) {
-    throw new Error("User not authenticated");
+  if (!userId) {
+    const {
+      data: { user },
+      error: authError,
+    } = await supabase.auth.getUser();
+    userId = user?.id;
+
+    if (authError || !userId) {
+      throw new Error("User not authenticated");
+    }
   }
 
   // isParticipating이 미제공(undefined)이고 관리자인 경우, 서버에서 직접 체크
